ima: add support for ima namespace
Now is possible to create Entry with lines on this template is inserted the logic
to support the verification. is possible to replicate the PCRs extension done by
childs. Is possible having just the list of the host to verificate its integrity.
The solution refers to the Implementation of the IMA namespace and extension of
the measure to parents.

Implementation of a client-verifier that is able to communicate to the cloud
verifier and ask for the verification of the whole host list and give back a list
regarding the namespace under attestation. The client verifier will check the
equailty between the list executed and golden values. The client verifer fiels
for configuration are inserted. The configuration is in the mapping.json file and
then dumped in clientverifier.js, the configuraitons are inserted in the
keylime.conf

Reference namespaces by container id in the APIs and perform polling on all the
agents because the position of the contianer is not know at priori. Now the API
to contact the agent will require the container id and will return the
identifier of the ima namespace and the list of measurements of the namespace.

Signed-off-by: Lorenzo Ferro <lorenzo.ferro@polito.it>
Signed-off-by: Enrico Bravi <enrico.bravi@polito.it>
Signed-off-by: Silvia Sisinni <silvia.sisinni@polito.it>